requirements.txt
autopep8==2.0.0
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2==2.9.5
pycodestyle==2.10.0
python-dotenv==0.21.0
SQLAlchemy==1.4.45
tomli==2.0.1
typing-extensions==4.4.0
Werkzeug==2.2.2
zipp==3.11.0
